Image_recognition/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
bzip2=1.0.6=3
cairo=1.14.12=h8948797_3
certifi=2016.2.28=py36_0
cffi=1.10.0=py36_0
cudatoolkit=8.0=3
cudnn=7.1.3=cuda8.0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=2.8.1=0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
graphite2=1.3.13=h23475e2_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
intel-openmp=2019.1=144
jasper=2.0.14=h07fcdf6_0
jbig=2.1=0
jpeg=9b=0
libedit=3.1=heed3624_0
libffi=3.2.1=1
libgcc=5.2.0=0
libgcc-ng=8.2.0=hdf63c60_1
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.36=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libuuid=1.0.3=0
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
mkl=2018.0.3=1
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
nccl=1.3.4=cuda8.0_1
ncurses=6.0=h9df7e31_2
ninja=1.7.2=0
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
olefile=0.44=py36_0
openssl=1.0.2l=0
pandas=0.20.3=py36_0
pcre=8.42=h439df22_0
pillow=5.4.1=py36h34e0f95_0
pip=9.0.1=py36_1
pixman=0.34.0=0
py-opencv=3.4.2=py36hb342d67_1
pycparser=2.18=py36_0
python=3.6.3=h0ef2715_3
python-dateutil=2.6.1=py36_0
pytorch=0.4.0=py36hd73e86b_0
pytz=2017.2=py36_0
readline=7.0=ha6073c6_4
scikit-learn=0.20.1=py36h4989274_0
scipy=1.1.0=py36hfa4b5c9_1
setuptools=36.4.0=py36_1
six=1.10.0=py36_0
sqlite=3.23.1=he433501_0
tk=8.6.8=hbc83047_0
torchvision=0.2.1=py36_0
tqdm=4.15.0=py36_0
wheel=0.29.0=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=0
